+++ _putchar.c
@@ -1,12 +1,10 @@
 #include <unistd.h>
 /**
- * _putchar - write the character c to stdout.
- * @c: The character to print.
- *
- * Return: On success, the function returns 1.
- * On error, the function returns -1 and errno is set.
+ * _putchar - write character c to the stdout
+ * @c: the character
+ * Return: 1 on success and -1 otherwise
  */
 int _putchar(char c)
 {
-return (write(1, &c, 1));
+	return (write(1, &c, 1));
 }
