+++ _strlen.c
@@ -1,0 +1,14 @@
+/**
+ * _strlen - length of a string
+ * @s: the string
+ * Return: the length
+ */
+
+int _strlen(char *s)
+{
+	int i;
+
+	for (i = 0; s[i] != '\0'; i++)
+		;
+	return (0);
+}
