+++ functions1.c
@@ -113,10 +113,10 @@ int print_hexa_upper(va_list types, char buffer[],
 		flags, 'X', width, precision, size));
 }
 
-/************** PRINT HEX (hexadecimal) NUM IN LOWER OR UPPER **************/
+/************** PRINT HEXX NUM IN LOWER OR UPPER **************/
 /**
- * print_hexa ;- prints hexadecimal number in lower or upper
- * @types: List of arguments
+ * print_hexa - Prints a hexadecimal number in lower or upper
+ * @types: Lista of arguments
  * @map_to: Array of values to map the number to
  * @buffer: Buffer array to handle print
  * @flags:  Calculates active flags
@@ -127,7 +127,6 @@ int print_hexa_upper(va_list types, char buffer[],
  * @size: Size specification
  * Return: Number of chars printed
  */
-
 int print_hexa(va_list types, char map_to[], char buffer[],
 	int flags, char flag_ch, int width, int precision, int size)
 {
