+++ functions2.c
@@ -1,4 +1,3 @@
-
 #include "main.h"
 
 /****************** PRINT POINTER ******************/
