+++ get_precision.c
@@ -8,7 +8,6 @@
  *
  * Return: Precision.
  */
-
 int get_precision(const char *format, int *i, va_list list)
 {
 	int curr_i = *i + 1;
