+++ get_size.c
@@ -7,7 +7,6 @@
  *
  * Return: Precision.
  */
-
 int get_size(const char *format, int *i)
 {
 	int curr_i = *i + 1;
