+++ get_width.c
@@ -8,7 +8,6 @@
  *
  * Return: width.
  */
-
 int get_width(const char *format, int *i, va_list list)
 {
 	int curr_i;
